requirements.txt
# Core Pydance  dependencies (minimal installation)
# Install with: pip install pydance 
uvloop>=0.15.0; sys_platform != 'win32'
httptools>=0.3.0
jinja2>=3.0.0
pydantic>=1.8.0
cryptography>=3.4.0
pyjwt>=2.0.0
bcrypt>=3.2.0
aiofiles>=0.8.0
python-multipart>=0.0.5
itsdangerous>=2.0.0
click>=8.0.0
rich>=10.0.0
typer>=0.4.0
email-validator>=1.1.0
python-dotenv>=0.19.0

# Essential features that are actually used
phonenumbers>=8.12.0
babel>=2.9.0
pytz>=2021.0
boto3>=1.20.0
aioredis>=2.0.0
aiohttp>=3.8.0
psutil>=5.8.0

# Optional dependencies for extended functionality:
# Database support: pip install pydance [database]
# Web3/Blockchain: pip install pydance [web3]
# Monitoring: pip install pydance [monitoring]
# Email features: pip install pydance [email]
# Storage: pip install pydance [storage]
# Microservices: pip install pydance [microservices]
# Streaming: pip install pydance [streaming]
# Documentation: pip install pydance [docs]
# Development: pip install pydance [dev]
# All features: pip install pydance [all]




